# 3-technical/3.1-system-foundation/architecture/api-contracts/openapi.yaml
openapi: 3.1.0
info:
  title: AKAIUNSAN Attendance + Payroll API
  version: 1.0.0
  description: |
    REST API for AKAIUNSAN attendance and payroll systems (PRD-EPIC-002, Phase 0).

    **Auth:** Bearer JWT (15-min TTL) + refresh token cookie (30 days).
    **Internal:** `X-Internal-API-Key` header (service-to-service only).
    **Rate limits:** Production API instances enforce a global 100 requests/minute limit; auth endpoints add OTP/challenge abuse controls.
    **Time:** ISO 8601 UTC; VN time (UTC+7) at presentation layer.
    **Money:** VNĐ as string (Decimal); never float.
    **Errors:** `{ error: { code, message, details? } }`

    Full schemas live in this file. Full business rules in
    [Domain Specs](../domain-specs.md). Architecture in
    [System Design](../../design-standards/system-design.md).

    Code conventions in [Coding Standards](../../design-standards/coding-standards.md).

  contact:
    name: AKAIUNSAN Engineering
    email: engineering@ak.local
...
security:
  - bearerAuth: []
paths:
...
  /v1/attendance/check-in:
    post:
      tags: [attendance]
      summary: Employee check-in (GPS + photo)
      description: |
        Enforces BR-ATT-001 (GPS in geofence), BR-ATT-002 (late), BR-ATT-004 (no double), BR-ATT-005 (photo required).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckInRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceRecord' }
        '400':
          description: Invalid request or non-JPEG photo data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/BusinessRuleViolation' }
  /v1/attendance/check-out:
    post:
      tags: [attendance]
      summary: Employee check-out (GPS + photo)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CheckOutRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceRecord' }
        '400':
          description: Invalid request or non-JPEG photo data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/BusinessRuleViolation' }
  /v1/attendance/records:
    get:
      tags: [attendance]
      summary: Query attendance records (admin/supervisor)
      parameters:
        - in: query
          name: employeeId
          schema: { type: string, format: uuid }
        - in: query
          name: projectId
          schema: { type: string, format: uuid }
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
        - in: query
          name: status
          schema: { type: string, enum: [present, late, early_leave, half_day, absent, on_leave, holiday] }
        - in: query
          name: limit
          description: Maximum records returned; defaults to 100.
          schema: { type: integer, minimum: 1, maximum: 500, default: 100 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/AttendanceRecordWithAssignment' }
  /v1/attendance/records/{id}/override:
    post:
      tags: [attendance]
      summary: Override attendance record (supervisor/admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AttendanceOverride' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AttendanceRecord' }

  # ===== CUSTOMER REPORTS =====
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token (15-min TTL)
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: Shared secret, rotated quarterly
...
  responses:
...
    BusinessRuleViolation:
      description: Domain rule violated (e.g., GPS outside geofence)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, additionalProperties: true }
...
    GPSCoordinate:
      type: object
      required: [latitude, longitude, accuracy]
      properties:
        latitude: { type: number, format: double, minimum: -90, maximum: 90 }
        longitude: { type: number, format: double, minimum: -180, maximum: 180 }
        accuracy: { type: number, format: double, minimum: 0, description: GPS accuracy in meters }

    # ===== Auth =====
...
    AttendanceEmployeeView:
      type: object
      required: [id, userId, employeeCode, fullName, status]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        employeeCode: { type: string }
        fullName: { type: string }
        status: { type: string, enum: [active, inactive] }
...
    Project:
      type: object
      required: [id, code, name, clientName, address, latitude, longitude, geofenceRadiusMeters, status]
      properties:
        id: { type: string, format: uuid }
        code: { type: string, example: PRJ001 }
        name: { type: string }
        clientName: { type: string }
        address: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        geofenceRadiusMeters: { type: integer, default: 100 }
        contractStartDate: { type: string, format: date }
        contractEndDate: { type: [string, 'null'], format: date }
        status: { type: string, enum: [active, paused, ended] }
...
    Shift:
      type: object
      required: [id, name, startTime, endTime, breakMinutes, lateThresholdMinutes, isOvernight]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: 'Ca sáng' }
        startTime: { type: string, example: '06:00' }
        endTime: { type: string, example: '14:00' }
        breakMinutes: { type: integer, default: 60 }
        lateThresholdMinutes: { type: integer, default: 15 }
        isOvernight: { type: boolean, default: false }
        color: { type: string, example: '#FF5733' }
...
    ShiftAssignment:
      type: object
      required: [id, employeeId, projectId, shiftId, date, status]
      properties:
        id: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        shiftId: { type: string, format: uuid }
        date: { type: string, format: date }
        status: { type: string, enum: [scheduled, checked_in, checked_out, completed, missed, cancelled] }
...
    AttendanceRecord:
      type: object
      required: [id, shiftAssignmentId, status]
      properties:
        id: { type: string, format: uuid }
        shiftAssignmentId: { type: string, format: uuid }
        employeeId: { type: string, format: uuid }
        projectId: { type: string, format: uuid }
        checkInAt: { type: [string, 'null'], format: date-time }
        checkOutAt: { type: [string, 'null'], format: date-time }
        checkInGps: { anyOf: [{ $ref: '#/components/schemas/GPSCoordinate' }, { type: 'null' }] }
        checkOutGps: { anyOf: [{ $ref: '#/components/schemas/GPSCoordinate' }, { type: 'null' }] }
        checkInPhotoUrl: { type: [string, 'null'] }
        checkOutPhotoUrl: { type: [string, 'null'] }
        status: { type: string, enum: [present, late, early_leave, half_day, absent, on_leave, holiday] }
        totalMinutesWorked: { type: [integer, 'null'] }
        overtimeMinutes: { type: [integer, 'null'] }
        lateMinutes: { type: [integer, 'null'] }
        overrideReason: { type: [string, 'null'] }
        overrideById: { type: [string, 'null'], format: uuid }
        overrideAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AttendanceRecordWithAssignment:
      allOf:
        - $ref: '#/components/schemas/AttendanceRecord'
        - type: object
          required: [shiftAssignment]
          properties:
            shiftAssignment:
              allOf:
                - $ref: '#/components/schemas/ShiftAssignment'
                - type: object
                  required: [employee, project, shift]
                  properties:
                    employee: { $ref: '#/components/schemas/AttendanceEmployeeView' }
                    project: { $ref: '#/components/schemas/Project' }
                    shift: { $ref: '#/components/schemas/Shift' }
...
    CheckInRequest:
      type: object
      required: [shiftAssignmentId, gps, photoBase64]
      properties:
        shiftAssignmentId: { type: string, format: uuid }
        gps: { $ref: '#/components/schemas/GPSCoordinate' }
        photoBase64:
          type: string
          description: JPEG base64-encoded, max 5 MB
    CheckOutRequest:
      type: object
      required: [shiftAssignmentId, gps, photoBase64]
      properties:
        shiftAssignmentId: { type: string, format: uuid }
        gps: { $ref: '#/components/schemas/GPSCoordinate' }
        photoBase64: { type: string }
    AttendanceOverride:
      type: object
      required: [reason, newStatus]
      properties:
        reason: { type: string, minLength: 10 }
        newStatus: { type: string, enum: [present, late, early_leave, half_day, absent, on_leave, holiday] }
        checkInAt: { type: [string, 'null'], format: date-time }
        checkOutAt: { type: [string, 'null'], format: date-time }
...